update1-wasd controls v1
This commit allows for the user to use the wasd keys to move the person around the screen. It is a little buggy, but functional. Need to figure out how to clear the Char ch after every event so that the bugs hopefully go away along with the long line of text which is the user input to move the character