working bush
this commit contains code for a single working bush. trying to use a looping system to spawn multiple bushes to avoid repetitive code. Also if I do not want to avoid using a looping system because without it I would have to have a specified number of objects(in this case bushes) that will be used every time the game is ran.